fast_backend/requirements.txt
aiofiles==23.1.0
anyio==3.6.2
asgiref==3.6.0
azure-core==1.26.4
azure-storage-blob==12.16.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.95.1
geographiclib==2.0
geopy==2.3.0
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
isodate==0.6.1
mango-jwt==1.3.2
Markdown==3.4.3
openpyxl==3.1.2
passlib==1.7.4
protobuf==4.23.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==3.12.3
pyroute2==0.7.8
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.2.4
starlette==0.26.1
tenacity==8.2.2
typing-extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
zipp==3.15.0
